be_visual_plot

Plot to compare barometric efficiency. Large datasets may take a long time to plot. Subsample should be set to TRUE

Usage

be_visual_plot(dat, time = "datetime", subsample = TRUE)

Arguments

dat

data that has the independent and dependent variables (data.table)

time

name of the column containing the time (character)

subsample

should the data be subsampled for plotting? (logical)

Value

plotly graph for barometric efficiency estimation with Smith method

Examples

library(plotly)
#> Loading required package: ggplot2
#> 
#> Attaching package: ‘plotly’
#> The following object is masked from ‘package:ggplot2’:
#> 
#>     last_plot
#> The following object is masked from ‘package:stats’:
#> 
#>     filter
#> The following object is masked from ‘package:graphics’:
#> 
#>     layout
library(data.table)
be <- 0.43
x <- seq(0, 28 * pi, pi / (12 * 12))

baro <- sin(x) + rnorm(length(x), sd = 0.04)
wl <- -sin(x) * be + rnorm(length(x), sd = 0.04)
dat <- data.table(datetime = as.POSIXct(x * 86400 / (2 * pi),
                                        origin = '1970-01-01', tz = 'UTC'),
                  wl = wl, baro = baro)
dat_be <- be_visual_data(dat)
#be_visual_plot(dat_be) #not run