step_add_vars

Add a variable from the initial (template) data not included in the recipe creation.

Usage

step_add_vars(.rec, terms, role = "predictor", ...)

Arguments

.rec

the R6 recipe object.

terms

the unquoted names of the variables to use or a selector function. terms replaces the `...` of the recipes package but requires variables to be included within `c()`. For example to include variables x and y you would write `c(x,y)` in the hydrorecipes package.

role

character - the name of the role

...

additional arguments

Value

recipe with added variables

Examples

dat <- data.frame(x = rnorm(10), y = rnorm(10), z = rnorm(10))

rec <- recipe(y~x, data = dat) |>
       step_add_vars(z) |>
       plate()