step_aquifer_grf

Generates the drawdown using the Generalized Radial Flow (GRF) model. This method defaults to a fast FFT convolution so many rates can be included, but requires a regular time series.

Usage

step_aquifer_grf(
  .rec,
  time,
  flow_rate,
  thickness = 1,
  radius = 100,
  specific_storage = 1e-06,
  hydraulic_conductivity = 1e-04,
  flow_dimension = 2,
  role = "predictor",
  ...
)

Arguments

.rec

the R6 recipe object.

time

the time for evaluation (t)

flow_rate

the flow rate from the well (L^3/t)

thickness

the aquifer thickness (L)

radius

the distance to the observation well (L)

specific_storage

specific storage of aquifer (L/L)

hydraulic_conductivity

the hydraulic conductivity (L/t)

flow_dimension

aquifer flow dimension (1 = linear, 2 = radial, 3 = spherical)

role

character - the name of the role

...

additional arguments

Value

The drawdown using the GRF model

References

Barker, J.A., A generalized radial flow model for hydraulic tests in fractured rock. Water Resour. Res., 24 (1988), pp. 1796-1804, 10.1029/WR024i010p01796

See also

Examples

time <- 1:2000
flow_rate  <- c(rep(0.001, 500),
                rep(0.002, 500),
                rep(0.0, 1000))

dat <- data.frame(time, flow_rate)

# radial (flow_dimension = 2 Theis)
dd_rad <- recipe(time~flow_rate, dat) |>
  step_aquifer_grf(time = time,
                   flow_rate = flow_rate,
                   thickness = 1.0,
                   radius = 20,
                   specific_storage = 1e-5,
                   hydraulic_conductivity = 1e-3,
                   flow_dimension = 2) |>
  plate()

# linear (flow_dimension = 1)

dd_lin <- recipe(time~flow_rate, dat) |>
  step_aquifer_grf(time = time,
                   flow_rate = flow_rate,
                   thickness = 1.0,
                   radius = 20,
                   specific_storage = 1e-5,
                   hydraulic_conductivity = 1e-3,
                   flow_dimension = 1) |>
  plate()

# spherical (flow_dimension = 3)
dd_sph <- recipe(time~flow_rate, dat) |>
  step_aquifer_grf(time = time,
                   flow_rate = flow_rate,
                   thickness = 1.0,
                   radius = 20,
                   specific_storage = 1e-5,
                   hydraulic_conductivity = 1e-3,
                   flow_dimension = 3) |>
  plate()