step_aquifer_theis_aniso

Generates the drawdown using the Papadopulos 1965 model. Papadopulos, I.S., 1965. Nonsteady flow to a well in an infinite anisotropic aquifer.

Usage

step_aquifer_theis_aniso(
  .rec,
  time,
  flow_rate,
  thickness = 1,
  distance_x = 100,
  distance_y = 100,
  specific_storage = 1e-06,
  hydraulic_conductivity_major = 1e-04,
  hydraulic_conductivity_minor = 1e-05,
  major_axis_angle = 0,
  role = "predictor",
  ...
)

Arguments

.rec

the R6 recipe object.

time

the time for evaluation (t)

flow_rate

the flow rate from the well (L^3/t)

thickness

the aquifer thickness (L)

distance_x

distance in the x direction

distance_y

distance in the y direction

specific_storage

specific storage of aquifer (L/L)

hydraulic_conductivity_major

hydraulic conductivity in the major principal direction

hydraulic_conductivity_minor

hydraulic conductivity in the minor principal direction

major_axis_angle

the orientation of the major principal axis (angle between the x axis and major axis)

role

character - the name of the role

...

additional arguments

Value

The drawdown using the Papadopulos 1965 model

References

Papadopulos, I.S., 1965. Nonsteady flow to a well in an infinite anisotropic aquifer. Heilweil, V.M. and Hsieh, P.A., 2006. Determining anisotropic transmissivity using a simplified Papadopulos method. Groundwater, 44(5), pp.749-753.

See also

Examples

dat <- data.frame(x = as.numeric(1:100),
                  y = rep(0.01, 100))
formula <- as.formula(y~x)

frec = recipe(formula = formula, data = dat) |>
  step_aquifer_theis_aniso(time = x, flow_rate = y) |>
  prep() |>
  bake()