step_baro_harmonic

Estimate the static barometric efficiency using harmonic methods (Ratio, Acworth, Rau, Transfer)

Usage

step_baro_harmonic(
  .rec,
  time,
  water_level,
  barometric_pressure,
  earth_tide,
  frequency = c(1.9324, 2),
  cycle_size = 86400,
  start = 0,
  inverse = FALSE,
  role = "augment",
  ...
)

Arguments

.rec

the R6 recipe object.

water_level

variable unquoted water level column name

barometric_pressure

variable unquoted barometric pressure column name

earth_tide

variable unquoted Earth tide column name

frequency

numeric vector - the frequencies of the sin and cos curves

cycle_size

numeric - the period of the sin and cos curves

inverse

logical whether the barometric relationship is inverse

role

character - the name of the role

...

additional arguments

Value

double barometric efficiency using different methods

References

Acworth, R.I., Halloran, L.J., Rau, G.C., Cuthbert, M.O. and Bernardi, T.L., 2016. An objective frequency domain method for quantifying confined aquifer compressible storage using Earth and atmospheric tides. Geophysical Research Letters, 43(22), pp.11-671.

See also

Examples

data("kennel_2020")
library(data.table)
library(collapse)
#> collapse 2.0.14, see ?`collapse-package` or ?`collapse-documentation`
#> 
#> Attaching package: ‘collapse’
#> The following object is masked from ‘package:data.table’:
#> 
#>     fdroplevels
#> The following object is masked from ‘package:stats’:
#> 
#>     D

formula <- as.formula(wl~.)
frec = recipe(formula = formula, data = kennel_2020) |>
 step_baro_harmonic(datetime,
                    wl,
                    baro,
                    et,
                    inverse = FALSE)