step_fft_transfer_welch

calculates the transfer function with Welch's results.

Usage

step_fft_transfer_welch(
  .rec,
  terms,
  length_subset,
  overlap = 0.8,
  window,
  time_step = 1,
  formula = NULL,
  role = "augment",
  ...
)

Arguments

.rec

the R6 recipe object.

terms

the unquoted names of the variables to use or a selector function. terms replaces the `...` of the recipes package but requires variables to be included within `c()`. For example to include variables x and y you would write `c(x,y)` in the hydrorecipes package.

length_subset

length of fft section

overlap

amount of overlap

window

window weights

time_step

numeric monitoring interval size

formula

formula notation to specify inputs and outputs

role

character - the name of the role

...

additional arguments

Value

an updated recipe

Examples

data(kennel_2020)
form <- as.formula("wl~.")

  rec <- recipe(form, kennel_2020) |>
  step_fft_transfer_welch(c(wl, baro, et),
                          length_subset = 1440*8 + 1,
                          overlap = 0.6,
                          window = window_nuttall(1440*8+1)) |>
  plate()