vadose_response

weeks_1979 1-D air diffusivity

Usage

vadose_response(time, air_diffusivity, thickness, precision, inverse)

Arguments

time

numeric vector of elapsed times

air_diffusivity

A numeric value of the unsaturated zone air diffusivity

thickness

A numeric value of the unsaturated zone thickness

precision

A numeric value of for the solution precision

inverse

A logical value indicating if an inverse water level relationship is desired

Value

weeks 1979 model

Examples

vr <- vadose_response(time = 0:43200,
                       air_diffusivity = 0.20,
                       thickness = 40,
                       precision = 1e-10,
                       inverse = FALSE)